Declare fields in PatchscopeBase
PatchscopeBase references the fields inside the child class Patchscope.

This breaks Intellisense. So, I'm declaring all fields inside PatchscopeBase.
To do this properly, I either need to move SourceContex & TargetContext to
patchscope_base.py, or move PatchscopeBase to patchscope.py. I picked the
later.
